Add a generic unit test helper for a language.
